[plugin,kodiext] added support for quick enigma2/kodi switching
kodiext - added K and E options,
 E - sends message to switch to enigma2 to UDS server and hides kodi
 K - sends message to switch to kodi to UDS server and shows kodi

plugin
- added resume support, in case kodi is already running
UDS server now runs from start of enigma2 until exit of enigma2. When
we start Kodi launcher, it checks if kodi is not running. If it's not
running we start kodi process as usual. If kodi is running we find out
what pid it is and use kodiext -K -P $kodipid command which sends
message to switch to kodi and shows kodi screen.

Note: It's also possible to run kodi, switch to enigma2, restart enigma2 and
again switch to kodi which is still running from last restart of enigma2

- added kodi to extensions menu (blue button) for quick access